@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {HttpClient} from "@angular/common/http";
 import {Router} from "@angular/router";
+import {AuthService} from "../../../services/auth.service";
 
 @Component({
   selector: 'login',
@@ -16,7 +16,7 @@ export class LoginComponent implements OnInit {
 
   showMessageError = false;
 
-  constructor(private http: HttpClient, private router: Router) {
+  constructor(private authService: AuthService, private router: Router) {
 
   }
 
@@ -25,10 +25,8 @@ export class LoginComponent implements OnInit {
 
     submit(){
         //generics - Java
-        this.http.post<any>('http://localhost:8000/api/login', this.credentials)
+        this.authService.login(this.credentials)
             .subscribe((data) => {
-                const token = data.token;
-                window.localStorage.setItem('token', token);
                 this.router.navigate(['categories/list']);
             }, () => this.showMessageError = true);
         return false;
