@@ -2,9 +2,10 @@ import { Injectable } from '@angular/core';
 import {HttpClient, HttpParams} from "@angular/common/http";
 import {Observable} from "rxjs/internal/Observable";
 import {Category} from "../../Models";
-import {urlApi} from "../../app.params";
 import {map} from "rxjs/operators";
 import {SearchParams, SearchParamsBuilder, HttpResource} from "./http-resource";
+import {AuthService} from "../auth.service";
+import {environment} from "../../../environments/environment";
 
 
 //singleton
@@ -14,76 +15,51 @@ import {SearchParams, SearchParamsBuilder, HttpResource} from "./http-resource";
 })
 export class CategoryHttpService {
 
-  private url : string = urlApi + 'categories';
+    private url : string = `${environment.api.url}/categories`;
 
   constructor(private http: HttpClient) {
 
   }
 
   list(searchParams: SearchParams): Observable<{data: Array<Category>, meta: any}>{
-      const token = window.localStorage.getItem('token');
       const sParams = new SearchParamsBuilder(searchParams).makeObject();
       const params = new HttpParams({
           fromObject: (<any>sParams)
       });
       return this.http
           .get<{data: Array<Category>, meta: any}>
-            (`${this.url}`, {
-                    params,
-                    headers:{
-                        'Authorization': `Bearer ${token}`
-                    }
-            });
+            (`${this.url}`, {params});
   }
 
   get(id: number): Observable<Category>{
-      const token = window.localStorage.getItem('token');
       return this.http
           .get<{ data: Category }>
-            (`${this.url}/${id}`, {
-          headers:{
-              'Authorization': `Bearer ${token}`
-          }
-      })
+            (`${this.url}/${id}`)
           .pipe(
               map(response => response.data)
           )
   }
 
   create(category: Category): Observable<Category>{
-      const token = window.localStorage.getItem('token');
-      return this.http.post<{data: Category}>
-        (this.url, category, {
-          headers:{
-              'Authorization': `Bearer ${token}`
-          }
-      })
+      return this.http
+          .post<{data: Category}>
+        (this.url, category)
           .pipe(
               map(response => response.data)
           )
   }
 
   update(id: number, category: Category): Observable<Category>{
-      const token = window.localStorage.getItem('token');
       return this.http
-          .put<{ data: Category }>(`${this.url}/${id}`, category, {
-          headers:{
-              'Authorization': `Bearer ${token}`
-          }
-      })
+          .put<{ data: Category }>(`${this.url}/${id}`, category)
           .pipe(
               map(response => response.data)
           )
   }
 
   destroy(id: number): Observable<any>{
-      const token = window.localStorage.getItem('token');
       return this.http
           .delete
-          (`${this.url}/${id}`, {
-              headers:{
-                  'Authorization': `Bearer ${token}`
-              }
-          })
+          (`${this.url}/${id}`)
   }
 }
