@@ -47,7 +47,7 @@ class User extends Authenticatable implements JWTSubject
     {
         return [
           'email' => $this->email,
- //         'name' => $this->name
+          'name' => $this->name
         ];
     }
 }
