@@ -21,9 +21,7 @@ Route::group(['namespace' => 'Api', 'as' => 'api.' ], function (){
     Route::name('login')->post('login', 'AuthController@login');
     Route::name('refresh')->post('refresh', 'AuthController@refresh');
 
-    Route::group(['middleware' => ['auth:api',
-    //    'jwt.refresh'
-    ]], function(){
+    Route::group(['middleware' => ['auth:api', 'jwt.refresh']], function(){
         Route::name('logout')->post('logout', 'AuthController@logout');
         Route::name('me')->get('me', 'AuthController@me');
         Route::patch('products/{product}/restore','ProductController@restore');
